refactor(#274): split SettingsPanel.tsx into per-tab components
Extract 8 inline tab components and 2 shared UI primitives from the
1439-line SettingsPanel.tsx into dedicated modules:

- GeneralTab, GitTab, ModelTab, WorkflowsTab, AppearanceTab,
  AdvancedTab, NotificationsTab, StatusTab
- SettingField, ToggleSwitch (shared UI)
- types.ts (shared TabProps interface)

SettingsPanel.tsx now serves as a thin orchestrator at 274 lines.
