+++ apps/desktop/src/components/settings/SettingField.tsx
@@ -1,0 +1,23 @@
+import type React from 'react';
+
+export function SettingField({
+  label,
+  description,
+  children,
+}: {
+  label: string;
+  description?: string;
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="space-y-2">
+      <div>
+        <label className="text-sm font-medium text-foreground">{label}</label>
+        {description && (
+          <p className="text-xs text-muted-foreground mt-0.5">{description}</p>
+        )}
+      </div>
+      {children}
+    </div>
+  );
+}
