+++ apps/desktop/src/components/settings/ToggleSwitch.tsx
@@ -1,0 +1,27 @@
+export function ToggleSwitch({
+  checked,
+  onChange,
+  'data-testid': testId,
+}: {
+  checked: boolean;
+  onChange: (checked: boolean) => void;
+  'data-testid'?: string;
+}) {
+  return (
+    <button
+      data-testid={testId}
+      role="switch"
+      aria-checked={checked}
+      onClick={() => onChange(!checked)}
+      className={`relative inline-flex h-5 w-9 shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors ${
+        checked ? 'bg-primary' : 'bg-input'
+      }`}
+    >
+      <span
+        className={`pointer-events-none inline-block h-4 w-4 rounded-full bg-background shadow-sm transition-transform ${
+          checked ? 'translate-x-4' : 'translate-x-0'
+        }`}
+      />
+    </button>
+  );
+}
