+++ apps/desktop/src/components/settings/types.ts
@@ -1,0 +1,6 @@
+import type { AppSettings } from '@/hooks/useSettings';
+
+export interface TabProps {
+  settings: AppSettings;
+  updateSettings: (updates: Partial<AppSettings>) => void;
+}
